first pass at refactor! buggy but running
